chore(duui-gnfinder-v2): Updated tests paths, artifactId in pom.xml
