+++ duui-gnfinder-v2/pom.xml
@@ -5,7 +5,7 @@
     <modelVersion>4.0.0</modelVersion>
 
     <groupId>org.texttechnologylab.duui</groupId>
-    <artifactId>GNFinder</artifactId>
+    <artifactId>duui-gnfinder-v2</artifactId>
     <version>2.0</version>
 
     <properties>
